refactor: move CLI commands to separate modules for better organization
- Create commands module with submodules for each command
- Move correct_last_epoch to commands/correct_epoch.rs
- Move send_message to commands/send_message.rs
- Move print_current_epoch to commands/current_epoch.rs
- Extract shared init_contracts helper to commands/mod.rs
- Clean up main.rs to only contain CLI parsing and dispatch
- This improves code organization and maintainability
